mincost bug mostly fixed, still ocurs infrequently

+++ blocks.cpp
@@ -21,7 +21,7 @@ Point rand_p(default_random_engine & eng){
 }
 void init_info(people & pps){
     default_random_engine eng(1);//uclock());
-    for(int i = 0; i < NUM_PEOPLE; i++){
+    for(size_t i = 0; i < NUM_PEOPLE; i++){
         Point home = rand_p(eng);
         uniform_int_distribution<int32_t> Xdist(max(0,home.X-int32_t(HOME_WORK_MAX_DIS)),min(int32_t(WORLD_SIZE-1),int32_t(home.X+HOME_WORK_MAX_DIS)));
         uniform_int_distribution<int32_t> Ydist(max(0,home.Y-int32_t(HOME_WORK_MAX_DIS)),min(int32_t(WORLD_SIZE-1),int32_t(home.Y+HOME_WORK_MAX_DIS)));
